+++ lib/utils/types/freezed/frosty_profile.dart
@@ -6,7 +6,7 @@ part 'frosty_profile.g.dart';
 
 @freezed
 class FrostyProfile with _$FrostyProfile {
-  factory FrostyProfile({required String name, required List<Mod> mods}) = _FrostyProfile;
+  factory FrostyProfile({required String name, required List<dynamic> mods}) = _FrostyProfile;
 
   factory FrostyProfile.fromJson(Map<String, dynamic> json) => _$FrostyProfileFromJson(json);
 }